requirements.txt
pandas>=2.0.0
psycopg2-binary>=2.9.5
jupyter>=1.0.0  # For running the debug.csv.ipynb notebook
notebook>=7.0.0  # Jupyter notebook server
requests>=2.28.0  # For making HTTP requests
